feat: postgres string data source adapter
