+++ packages/amplify-category-api/src/commands/api/import.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { printer } from '@aws-amplify/amplify-prompts';
 import fs from 'fs-extra';
-import { RDS_SCHEMA_FILE_NAME } from '@aws-amplify/graphql-transformer-core';
+import { RDS_SCHEMA_FILE_NAME, ImportedRDSType } from '@aws-amplify/graphql-transformer-core';
 import { importAppSyncAPIWalkthrough } from '../../provider-utils/awscloudformation/service-walkthroughs/import-appsync-api-walkthrough';
 import { getAPIResourceDir } from '../../provider-utils/awscloudformation/utils/amplify-meta-utils';
 import { writeSchemaFile, generateRDSSchema } from '../../provider-utils/awscloudformation/utils/graphql-schema-utils';
@@ -14,7 +14,19 @@ export const name = subcommand;
 
 export const run = async (context: $TSContext) => {
   printer.warn(PREVIEW_BANNER);
-  const importAppSyncAPIWalkInputs = await importAppSyncAPIWalkthrough(context);
+  // const importAppSyncAPIWalkInputs = await importAppSyncAPIWalkthrough(context);
+  // console.log(JSON.stringify(importAppSyncAPIWalkInputs, null, 2));
+  const importAppSyncAPIWalkInputs = {
+    apiName: 'rdsauthpg',
+    dataSourceConfig: {
+      engine: 'postgres' as ImportedRDSType,
+      database: 'testdb',
+      host: 'database-pg.cdsc49uwsz3f.us-east-1.rds.amazonaws.com',
+      port: 5432,
+      username: 'postgres',
+      password: 'KyYe4ZwqGi9Wc-*kPLLbx-',
+    },
+  };
 
   if (importAppSyncAPIWalkInputs?.dataSourceConfig) {
     // ensure imported API resource artifacts are created
