+++ packages/amplify-graphql-schema-generator/src/datasource-adapter/index.ts
@@ -3,3 +3,4 @@ export * from './string-datasource-adapter';
 export * from './mysql-datasource-adapter';
 export * from './pg-datasource-adapter';
 export { MySQLStringDataSourceAdapter } from './mysql-string-datasource-adapter';
+export { PostgresStringDataSourceAdapter } from './pg-string-datasource-adapter';
