+++ packages/amplify-graphql-schema-generator/src/datasource-adapter/string-datasource-adapter.ts
@@ -1,4 +1,6 @@
 import { parse } from 'csv-parse/sync';
+import { singular } from 'pluralize';
+import { toPascalCase } from 'graphql-transformer-common';
 import { Field, FieldType, Index, Model } from '../schema-representation';
 
 export abstract class StringDataSourceAdapter {
@@ -55,5 +57,9 @@ export abstract class StringDataSourceAdapter {
     primaryKey && model.setPrimaryKey(primaryKey.getFields());
     indexes.forEach((index) => model.addIndex(index.name, index.getFields()));
     return model;
+  }
+
+  protected getEnumName(name: string): string {
+    return singular(toPascalCase(name.split('_')));
   }
 }
